get unit testing working
