+++ .gitignore
@@ -1,5 +1,7 @@
-app/cache/*
-app/logs/*
-composer.lock
+# IDE generated files
+.idea
+
+# Application files
+/vendor
 composer.phar
-vendor/*
+composer.lock
