+++ ObjectTranslationBundle.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ObjectTranslationBundle;
+namespace Skippy565\ObjectTranslationBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
