+++ README.md
@@ -1,38 +1,34 @@
-Installation
-============
+# Installation
 
-Step 1: Download the Bundle
----------------------------
+## Step 1: Download the Bundle
 
 Add the following to your project's composer.json file:
 
-.. code-block:: json
-
+```
     "repositories": [
         {
             "type": "vcs",
             "url": "git@github.com:Skippy565/ObjectTranslationBundle.git"
         }
     ]
+```
 
 Open a command console, enter your project directory and execute the
 following command to download the latest stable version of this bundle:
 
-.. code-block:: bash
-
+```bash
     $ composer require Skippy565/ObjectTranslationBundle:1.0.*
+```
 
 This command requires you to have Composer installed globally, as explained
-in the `installation chapter`_ of the Composer documentation.
+in the [installation chapter](https://getcomposer.org/doc/00-intro.md) of the Composer documentation.
 
-Step 2: Enable the Bundle
--------------------------
+## Step 2: Enable the Bundle
 
 Then, enable the bundle by adding it to the list of registered bundles
-in the ``app/AppKernel.php`` file of your project:
+in the `app/AppKernel.php` file of your project:
 
-.. code-block:: php
-
+```php
     <?php
     // app/AppKernel.php
 
@@ -52,5 +48,4 @@ in the ``app/AppKernel.php`` file of your project:
 
         // ...
     }
-
-.. _`installation chapter`: https://getcomposer.org/doc/00-intro.md
+```
