--- Tests/Controller/TestToObject.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace ObjectTranslationBundle\Tests\Controller;
-
-class TestToObject
-{
-	protected $fullName;
-	protected $email;
-}
