+++ Tests/Controller/ToTestObject.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Skippy565\ObjectTranslationBundle\Tests\Controller;
+
+/**
+ * Class ToTestObject
+ * @package Skippy565\ObjectTranslationBundle\Tests\Controller
+ */
+class ToTestObject
+{
+    public $firstName;
+
+    public $lastName;
+
+    public $email;
+}
