+++ phpunit.xml.dist
@@ -13,7 +13,7 @@
 
     <testsuites>
         <testsuite name="Project Test Suite">
-            <directory>../Tests</directory>
+            <directory>Tests/*</directory>
         </testsuite>
     </testsuites>
 </phpunit>
